requirements.txt
absl-py==2.1.0
addict==2.4.0
albucore==0.0.23
albumentations==2.0.4
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
asynctest==0.13.0
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
black==25.1.0
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.1
ccimport==0.4.4
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.8
codecov==2.1.13
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.1
coverage==7.6.11
crcmod==1.7
cryptography==44.0.0
cumm==0.7.11
cycler==0.12.1
Cython==3.0.11
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
descartes==1.1.0
docker-pycreds==0.4.0
easydict==1.13
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.14.0
fire==0.7.0
flake8==7.1.1
Flask==3.0.3
flatbuffers==25.1.24
fonttools==4.55.8
fqdn==1.5.1
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.70.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.4
imageio==2.37.0
importlib_metadata==8.6.1
iniconfig==2.0.0
interrogate==1.7.0
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isoduration==20.11.0
isort==6.0.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.8.0
kiwisolver==1.4.8
kwarray==0.7.1
lark==1.2.2
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.44.0
lyft-dataset-sdk==0.0.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.1
ml-dtypes==0.4.1
mmcv==2.1.0
mmdet==3.2.0
-e git+https://github.com/open-mmlab/mmdetection3d.git@fe25f7a51d36e3702f961e198894580d83c4387b#egg=mmdet3d
mmengine==0.10.3
model-index==0.1.11
mpmath==1.3.0
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.25.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
notebook==7.3.2
notebook_shim==0.2.4
numba==0.61.0
numpy==1.24.4
nuscenes-devkit==1.1.9
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
open3d==0.19.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
opt_einsum==3.4.0
optree==0.14.0
ordered-set==4.1.0
oss2==2.17.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pccm==0.4.16
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
plotly==6.0.0
pluggy==1.5.0
plyfile==0.7.4
pointpillars @ file:///mnt/disk2/users/milad/Research/3D-Bboxes/Pipeline_2/models/PointPillars
portalocker==3.1.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==3.20.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycocotools==2.0.8
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.1
pyparsing==3.2.1
pyquaternion==0.9.9
pytest==8.3.4
pytest-cov==6.0.0
pytest-runner==6.0.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2023.4
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scikit-image==0.19.3
scikit-learn==1.6.1
scipy==1.15.1
Send2Trash==1.8.3
sentry-sdk==2.21.0
setproctitle==1.3.4
Shapely==1.8.5.post1
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
spconv==2.3.8
stack-data==0.6.3
stringzilla==3.11.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
terminado==0.18.1
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2025.1.10
timm==1.0.14
tinycss2==1.4.0
torch==2.2.2+cu121
torchaudio==2.2.2+cu121
torchvision==0.17.2+cu121
tornado==6.4.2
tqdm==4.65.2
traitlets==5.14.3
trimesh==2.35.39
triton==2.2.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
ubelt==1.3.7
uri-template==1.3.0
urllib3==1.26.13
wandb==0.19.6
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.17.2
xdoctest==1.2.0
yapf==0.43.0
zipp==3.21.0
